reduce refcount from 64bit to 32bit

+++ bucket_map/src/lib.rs
@@ -8,4 +8,4 @@ mod bucket_storage;
 mod index_entry;
 
 pub type MaxSearch = u8;
-pub type RefCount = u64;
+pub type RefCount = u32;
